+++ travail/ProjetTutore1/Block.h
@@ -4,8 +4,8 @@
 #include "Player.h"
 class Block {
 private:
-	int _x = 0;
-	int _y = 0;
+	float _x = 0;
+	float _y = 0;
 	sf::IntRect *_rect;
 	sf::Sprite *_sprite;
 	sf::Texture* _tex;
@@ -14,13 +14,13 @@ protected:
 public :
 	//NEVER USE FOR DEBUG ONLY
 	bool colliding = false;
-	Block(int x, int y, sf::Image*);
+	Block(float x, float y, sf::Image*);
 	Block(Block *b);
 	~Block();
 	void show(sf::RenderWindow* rw);
 	bool collide(sf::IntRect rect);
-	inline int getX() { return _x; };
-	inline int getY() { return _y; };
+	inline float getX() { return _x; };
+	inline float getY() { return _y; };
 	friend bool operator==(const Block& b1, const Block& b2);
 	virtual bool effectPlayer(Player* p);
 };
